Fixed publish_customized_test_cycle Added zephyr token verification Replaced logging with print

+++ zephyr_results_publisher/publisher.py
@@ -1,19 +1,18 @@
 import json
-import logging
 import requests
 from zephyr_results_publisher.behave_to_cucumber_converter import convert_report, validate_json
 from zephyr_results_publisher.file_util import zip_file, get_path_dir
 from zephyr_results_publisher.helper import check_response_status, find_folder_id_by_name
 
-logging.basicConfig(level=logging.INFO, format='%(asctime)s :: %(levelname)s :: %(message)s')
-
 BASE_URL = "https://api.zephyrscale.smartbear.com/v2"
 
 
 def publish(zephyr_token, project_key, source_report_file, report_format, auto_create_test_cases="true"):
+    if does_zephyr_token_exist(zephyr_token) is False:
+        return
     if report_format == "behave":
         report_format = "cucumber"
-        logging.info(f"Start converting: {source_report_file}")
+        print(f"Start converting: {source_report_file}")
         report = convert_behave_report(source_report_file)
         if report is not None:
             with open(source_report_file, "w") as f:
@@ -37,11 +36,11 @@ def publish(zephyr_token, project_key, source_report_file, report_format, auto_c
         "file": open(output_zip, 'rb')
     }
 
-    logging.info(f"Sending results to Zephyr Scale...")
+    print(f"Sending results to Zephyr Scale...")
     response = requests.post(url, files=files, params=params, headers=headers)
     check_response_status(response, 200)
     parsed_response = json.loads(response.text)
-    logging.info(f"Parsed response: {parsed_response}")
+    print(f"Parsed response: {parsed_response}")
     return parsed_response
 
 
@@ -55,7 +54,11 @@ def publish_customized_test_cycle(zephyr_token,
                                   test_cycle_description="",
                                   test_cycle_jira_project_version=1,
                                   test_cycle_custom_fields=None):
-    if test_cycle_custom_fields is None:
+    if does_zephyr_token_exist(zephyr_token) is False:
+        return
+    if test_cycle_custom_fields is None \
+            or test_cycle_custom_fields == "" \
+            or test_cycle_custom_fields == "{}":
         test_cycle_custom_fields = {}
 
     if report_format == "behave":
@@ -91,12 +94,19 @@ def publish_customized_test_cycle(zephyr_token,
         "testCycle": ("test_cycle.json", test_cycle, "application/json")
     }
 
-    logging.info(f"Sending results to Zephyr Scale...")
+    print(f"Sending results to Zephyr Scale...")
     response = requests.post(url, files=files, params=params, headers=headers)
     check_response_status(response, 200)
     parsed_response = json.loads(response.text)
-    logging.info(f"Parsed response: {parsed_response}")
+    print(f"Parsed response: {parsed_response}")
     return parsed_response
+
+
+def does_zephyr_token_exist(zephyr_token):
+    if zephyr_token is None:
+        print("Zephyr Scale API Access Token is not provided. Results will not be published to Zephyr Scale.")
+        return False
+    return True
 
 
 def convert_behave_report(report_path):
@@ -128,7 +138,7 @@ def customize_test_cycle(zephyr_token,
         "folderId": folder_id,
         "customFields": custom_fields
     }
-    logging.info(f"Custom test cycle is generated: {test_cycle_json}")
+    print(f"Custom test cycle is generated: {test_cycle_json}")
     return json.dumps(test_cycle_json)
 
 
@@ -144,8 +154,6 @@ def get_folder_id_by_name(zephyr_token, name, project_key, max_results):
     headers = {
         "Authorization": "Bearer " + zephyr_token
     }
-
-    logging.info(params)
 
     response = requests.get(url, params=params, headers=headers)
     check_response_status(response, 200)
